+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Changed
-- Added `MmapOptions::no_reserve` method to support not to reserve swap space for this mapping. Linux only.
+- Added `MmapOptions::no_reserve` method to support not to reserve memory page or swap space for this mapping.
 
 ## [0.9.4] - 2024-01-25
 ### Changed